requirements.txt
django
djangorestframework
django-webpack-loader
markdown
django-filter
numpy
pandas
